Move uri handler, and change when clauses
